Add files via upload
In conclusion, I can say we have n sets of random sets of the data. n is just a random subset of the acquired dataset. Even, 100 is not a limit, I created more than 100 subsets such as a max of 1000 as well. so, ignore the fact 42 is a universal number.